requirements.txt
discord.py==1.7.3
pytz
pillow
requests
googletrans==4.0.0-rc1
sympy
apscheduler
tweepy
colorama